+++ src/jtl/Connector/Shopware/Mapper/Product.php
@@ -186,6 +186,7 @@ class Product extends DataMapper
         //$result = new ProductModel();
         $result = $product;
 
+        /*
         Logger::write(sprintf('>>> Product with id (%s, %s), masterProductId (%s, %s), manufacturerId (%s, %s)',
             $product->getId()->getEndpoint(),
             $product->getId()->getHost(),
@@ -194,6 +195,7 @@ class Product extends DataMapper
             $product->getManufacturerId()->getEndpoint(),
             $product->getManufacturerId()->getHost()
         ), Logger::DEBUG, 'database');
+        */
 
         try {
             if ($this->isChild($product)) {
@@ -635,7 +637,12 @@ class Product extends DataMapper
                 $i++;
                 foreach ($attribute->getI18ns() as $attributeI18n) {
                     if ($attributeI18n->getLanguageISO() === LanguageUtil::map(Shopware()->Shop()->getLocale()->getLocale())) {
-                        //$setter = 'set' . ucfirst($attributeI18n->getName());
+
+                        // Work Around, thx @db structure
+                        if ($i == 17) {
+                            $i++;
+                        }
+
                         $setter = "setAttr{$i}";
 
                         // active